armv6l bugfixes + less flickering status

+++ protocol/mpdu.h
@@ -38,9 +38,9 @@ inline uint8_t  mpdu_seq_flag(Mpdu *m) { return m->seq[0] >> 6; }
 inline uint16_t mpdu_seq(Mpdu *m) { return (m->seq[0] & 0x3F) << 8 | m->seq[1]; }
 inline uint16_t mpdu_len(Mpdu *m) { return (m->len[0] << 8 | m->len[1]) + 1; }
 inline uint16_t mpdu_day(Mpdu *m) { return m->data.time.day[0] << 8 | m->data.time.day[1]; }
-inline uint32_t mpdu_ms(Mpdu *m) { return m->data.time.ms[0] << 24 | m->data.time.ms[1] << 16 | m->data.time.ms[2] << 8 | m->data.time.ms[3]; }
+inline uint32_t mpdu_ms(Mpdu *m) { return (uint32_t)m->data.time.ms[0] << 24 | (uint32_t)m->data.time.ms[1] << 16 | (uint32_t)m->data.time.ms[2] << 8 | (uint32_t)m->data.time.ms[3]; }
 inline uint16_t mpdu_us(Mpdu *m) { return m->data.time.us[0] << 8 | m->data.time.us[1]; }
-inline uint64_t mpdu_raw_time(Mpdu *m) { return (uint64_t)mpdu_day(m)*86400L*1000L*1000L + mpdu_ms(m)*1000L + mpdu_us(m); }
+inline uint64_t mpdu_raw_time(Mpdu *m) { return (uint64_t)mpdu_day(m)*86400LL*1000LL*1000LL + (uint64_t)mpdu_ms(m)*1000L + (uint64_t)mpdu_us(m); }
 
 char *mpdu_time(uint64_t us);
 #endif
